WIP added Update command and Flow logic for branch flow, integrationTest and Merge logic todo
